+++ Touch_Emulator/src/main/java/excel/XLSXService.java
@@ -38,11 +38,11 @@ public interface XLSXService {
     
     String[] getProductNames(int day, int group, int subgroup);
     
-    int[] getProductPlu(int day, int group, int subgroup);
+    long[] getProductPlu(int day, int group, int subgroup);
     
     String readProductName(int day, int group, int subgroup, int product);
     
-    int readProductPlu(int day, int group, int subgroup, int product);
+    long readProductPlu(int day, int group, int subgroup, int product);
     
     void close();
 }
